requirements.txt
requests
python-dotenv
feedparser
pandas
matplotlib
reportlab
nltk
yfinance
